Add host execution profile events
Introduce a HAL profile host-execution event family for queue operation spans measured in IREE host time. The schema now has explicit relationship endpoint/type vocabulary for linking host queue events or submissions to host execution spans, and the tooling summary/model/projection/export paths preserve those spans separately from device-tick dispatch and queue-device events.

Bump the ireeperf-jsonl interchange schema to v4, emit host_execution_event rows, and teach iree-profile-render to accept v4, render host execution lanes, and consume both queue-submission and queue-event relationship flows. Keep Perfetto track ordering deterministic by spacing queue-device, dispatch, host-execution, memory, and counter track ranks.

AMDGPU command-buffer replay now stores the session-local executable id as uint64_t so recorded dispatch attribution matches the profile metadata namespace without a truncating side constraint.

+++ runtime/src/iree/hal/drivers/amdgpu/aql_command_buffer.c
@@ -2000,7 +2000,7 @@ static iree_status_t iree_hal_amdgpu_aql_command_buffer_dispatch(
   dispatch_command->group_segment_size =
       kernel_args->group_segment_size + config.dynamic_workgroup_local_memory;
   dispatch_command->executable_id =
-      (uint32_t)iree_hal_amdgpu_executable_profile_id(executable);
+      iree_hal_amdgpu_executable_profile_id(executable);
 
   if (binding_sources && !uses_custom_direct_arguments) {
     IREE_RETURN_IF_ERROR(
